Extract app:ui:audit and app:ui:components
Relocate shared widgets to keep layering acyclic:
- ScreenSizeClass -> foundation/util
- EditableAttributesSection, AttributeTypeField -> components
- CreateCategoryDialog, ParentCategorySelector -> components
- ErrorMessageText made public
