+++ app/ui/components/src/commonMain/kotlin/com/moneymanager/ui/components/DatabaseStartupProgressScreen.kt
