+++ app/ui/core/build.gradle.kts
@@ -19,6 +19,8 @@ kotlin {
                 api(projects.app.qifimporter)
                 api(projects.app.remotestorage.core)
                 api(projects.app.remotestorage.sync)
+                api(projects.app.ui.audit)
+                api(projects.app.ui.components)
                 api(projects.app.ui.foundation)
                 api(projects.utils.bigdecimal)
                 api(projects.utils.localsettings)
