+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/AccountsScreen.kt
@@ -743,118 +743,3 @@ fun UnmergeAccountDialog(
         },
     )
 }
-
-@Composable
-fun CreateCategoryDialog(
-    categoryRepository: CategoryReadRepository,
-    onCategoryCreated: (id: Long, name: String) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    var name by remember { mutableStateOf("") }
-    var selectedParentId by remember { mutableStateOf<Long?>(null) }
-    var expanded by remember { mutableStateOf(false) }
-    var errorMessage by remember { mutableStateOf<String?>(null) }
-    var isSaving by remember { mutableStateOf(false) }
-    val source = Source.Manual
-    val importEngine = LocalImportEngine.current
-
-    val categories by categoryRepository
-        .getAllCategories()
-        .collectAsStateWithSchemaErrorHandling(initial = emptyList())
-    val scope = rememberSchemaAwareCoroutineScope()
-
-    AlertDialog(
-        onDismissRequest = { if (!isSaving) onDismiss() },
-        title = { Text("Create New Category") },
-        text = {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 8.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp),
-            ) {
-                OutlinedTextField(
-                    value = name,
-                    onValueChange = { name = it },
-                    label = { Text("Category Name") },
-                    modifier = Modifier.fillMaxWidth(),
-                    singleLine = true,
-                    enabled = !isSaving,
-                )
-
-                ParentCategorySelector(
-                    categories = categories,
-                    selectedParentId = selectedParentId,
-                    onParentSelected = { selectedParentId = it },
-                    expanded = expanded,
-                    onExpandedChange = { expanded = it },
-                    enabled = !isSaving,
-                    additionalExcludedIds = emptySet(),
-                )
-
-                errorMessage?.let { error ->
-                    Text(
-                        text = error,
-                        color = MaterialTheme.colorScheme.error,
-                        style = MaterialTheme.typography.bodySmall,
-                    )
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(
-                onClick = {
-                    if (name.isBlank()) {
-                        errorMessage = "Category name is required"
-                    } else {
-                        isSaving = true
-                        errorMessage = null
-                        scope.launch {
-                            try {
-                                val key = LocalCategoryKey("create")
-                                val result =
-                                    importEngine.import(
-                                        ImportBatch.manualEdits(
-                                            categories =
-                                                listOf(
-                                                    ImportCategoryIntent(
-                                                        key = key,
-                                                        source = source,
-                                                        name = name.trim(),
-                                                        parentId = selectedParentId,
-                                                    ),
-                                                ),
-                                        ),
-                                    )
-                                onCategoryCreated(result.createdCategoryIds.getValue(key), name.trim())
-                            } catch (expected: Exception) {
-                                logger.error(expected) { "Failed to create category: ${expected.message}" }
-                                errorMessage = "Failed to create category: ${expected.message}"
-                                isSaving = false
-                            }
-                        }
-                    }
-                },
-                enabled = !isSaving,
-            ) {
-                if (isSaving) {
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(16.dp),
-                        strokeWidth = 2.dp,
-                    )
-                } else {
-                    Text("Create")
-                }
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = onDismiss,
-                enabled = !isSaving,
-            ) {
-                Text("Cancel")
-            }
-        },
-    )
-}
