+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/CategoriesScreen.kt
@@ -786,76 +786,6 @@ private fun saveCategoryWithValidation(
 }
 
 @Composable
-internal fun ParentCategorySelector(
-    categories: List<Category>,
-    selectedParentId: Long?,
-    onParentSelected: (Long?) -> Unit,
-    expanded: Boolean,
-    onExpandedChange: (Boolean) -> Unit,
-    enabled: Boolean,
-    additionalExcludedIds: Set<Long>,
-) {
-    var searchQuery by remember { mutableStateOf("") }
-    val filteredCategories =
-        remember(categories, searchQuery, additionalExcludedIds) {
-            val available =
-                categories.filter {
-                    it.id != Category.UNCATEGORIZED_ID && it.id !in additionalExcludedIds
-                }
-            if (searchQuery.isBlank()) available else available.filter { it.name.contains(searchQuery, ignoreCase = true) }
-        }
-
-    ExposedDropdownMenuBox(
-        expanded = expanded,
-        onExpandedChange = { onExpandedChange(!expanded && enabled) },
-    ) {
-        OutlinedTextField(
-            value =
-                if (expanded) {
-                    searchQuery
-                } else if (selectedParentId == null) {
-                    "None (Top Level)"
-                } else {
-                    categories.find { it.id == selectedParentId }?.name ?: "None (Top Level)"
-                },
-            onValueChange = { searchQuery = it },
-            label = { Text("Parent Category") },
-            placeholder = { Text("Type to search...") },
-            trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) },
-            modifier = Modifier.fillMaxWidth().menuAnchor(ExposedDropdownMenuAnchorType.PrimaryEditable),
-            enabled = enabled,
-            singleLine = true,
-        )
-        ExposedDropdownMenu(
-            expanded = expanded,
-            onDismissRequest = {
-                onExpandedChange(false)
-                searchQuery = ""
-            },
-        ) {
-            DropdownMenuItem(
-                text = { Text("None (Top Level)") },
-                onClick = {
-                    onParentSelected(null)
-                    onExpandedChange(false)
-                    searchQuery = ""
-                },
-            )
-            filteredCategories.forEach { category ->
-                DropdownMenuItem(
-                    text = { Text(category.name) },
-                    onClick = {
-                        onParentSelected(category.id)
-                        onExpandedChange(false)
-                        searchQuery = ""
-                    },
-                )
-            }
-        }
-    }
-}
-
-@Composable
 fun DeleteCategoryDialog(
     category: Category,
     onDismiss: () -> Unit,
