+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/transactions/TransactionCard.kt
@@ -1,5 +1,7 @@
 package com.moneymanager.ui.screens.transactions
 
+import com.moneymanager.ui.util.ScreenSizeClass
+
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
