+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/transactions/TransactionsScreen.kt
@@ -2,6 +2,8 @@
 
 package com.moneymanager.ui.screens.transactions
 
+import com.moneymanager.ui.util.ScreenSizeClass
+
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.horizontalScroll
